Refactored Registry to Locator.
